backend/pyproject.toml
[tool.poetry]
name = "modak"
version = "0.1.0"
description = "FastAPI backend for the Project Modak"
authors = ["Jooeon Kim <joo2theeon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
SQLAlchemy = "^1.4.40"
PyYAML = "^6.0"
pydantic = "^1.9.2"
python-dotenv = "^0.20.0"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.3"
uvicorn = "^0.18.2"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
